+++ lib/tortoise/handler/logger.ex
@@ -8,10 +8,12 @@ defmodule Tortoise.Handler.Logger do
 
   @behaviour Tortoise.Handler
 
+  @impl true
   def init(_opts) do
     {:ok, %State{}}
   end
 
+  @impl true
   def connection(:up, state) do
     Logger.info("Connection has been established")
     {:ok, state}
@@ -22,6 +24,7 @@ defmodule Tortoise.Handler.Logger do
     {:ok, state}
   end
 
+  @impl true
   def subscription(:up, topic, state) do
     Logger.info("Subscribed to #{topic}")
     {:ok, state}
@@ -42,11 +45,13 @@ defmodule Tortoise.Handler.Logger do
     {:ok, state}
   end
 
+  @impl true
   def handle_message(topic, publish, state) do
     Logger.info("#{Enum.join(topic, "/")} #{inspect(publish)}")
     {:ok, state}
   end
 
+  @impl true
   def terminate(reason, _state) do
     Logger.warn("Client has been terminated with reason: #{inspect(reason)}")
     :ok